requirements.txt
cfgv==3.4.0
distlib==0.3.9
filelock==3.16.1
identify==2.6.5
markdown-it-py==3.0.0
mdurl==0.1.2
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
pyaes==1.6.1
pyasn1==0.6.1
Pygments==2.19.1
python-dotenv==1.0.0
PyYAML==6.0.2
rich==13.7.0
rsa==4.9
Telethon==1.34.0
virtualenv==20.28.1
